Crate muoxi_staging

Main Proxy Staging TCP Server

This is where all clients will eventually connect to either via direct connection or one of the supported proxy servers (MCCP, Webserver, etc..)

Modules

cmds

comms
Definitions and declarations of data structures relating comms

copyover
A custom reimplementation of tokio::io::copy

prelude
Definitions for CommandSets. Depending on a variety of factors, you have access to different sets of commands. Some of the basic conditions ruling this would be:

states
Holds the different connection states for connected clients

Macros

cmdset
an easier way to create a command set from valid impl Command objects

Functions

client_cleanup
clean up client on disconnect or timeout

display_welcome
display welcome screen

get
Friendly async wrapper around recieving message from client Instead of panicing on wrong error, it will return an Option

main

process
Main processing piece of logic, once a connection is established to client the entire lifetime of the connected client is handled within this function.

send
Friendly async wrapper to sending messages to client object

transfer
Turns the staging server into a full proxy server, relaying information sent to proxy/staging server to MuOxi game engine