Function muoxi_staging::get

pub async fn get<'a>(client: &'a mut Client) -> String

Friendly async wrapper around recieving message from client Instead of panicing on wrong error, it will return a client disconnect message

let client_response = get(&client).await;
println!("Recieved from client: {}", client_response);