Module muoxi_staging::prelude

Definitions for CommandSets. Depending on a variety of factors, you have access to different sets of commands. Some of the basic conditions ruling this would be:

Connection State

Roles

Structs

CmdSet
Defines a common collection of commands

Statics

GAME_ADDR
Current listening port of the MuOxi game engine

PROXY_ADDR
Current listening port of the staging proxy server

Traits

Command
defines a command trait

Type Definitions

CommandResult
Custom error type revolving around the success of executing commands

LinesCodecResult
Result generic resulting with decoding/encoding errors

Rx
alias for recieving channel

Tx
alias for sending channel